mysite/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
jws==0.1.3
newsapi-python==0.2.6
oauth2client==3.0.0
pkg-resources==0.0.0
protobuf==3.12.0
pusher-push-notifications==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
PyJWT==1.7.1
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
